+++ src/rer/customersatisfaction/browser/show_feedbacks.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
+from AccessControl.unauthorized import Unauthorized
 from plone import api
+from Products.CMFPlone.PloneBatch import Batch
 from Products.Five import BrowserView
+from rer.customersatisfaction import _
 from rer.customersatisfaction.interfaces import ICustomerSatisfactionStore
 from zope.component import getUtility
-from Products.CMFPlone.PloneBatch import Batch
-from rer.customersatisfaction import _
-from plone.memoize import view
+from zope.publisher.interfaces import NotFound
 
 import logging
 
@@ -13,19 +14,42 @@ logger = logging.getLogger(__name__)
 
 
 class View(BrowserView):
-    @view.memoize
+    def __call__(self):
+        uid = self.request.form.get("uid", "")
+        if not uid:
+            api.portal.show_message(
+                message=_(
+                    "show_feedbacks_missing_uid",
+                    default=u"You need to provide a UID.",
+                ),
+                request=self.request,
+                type="error",
+            )
+            return self.request.response.redirect(api.portal.get().portal_url())
+        self.check_access(uid=uid)
+        return super(View, self).__call__()
+
+    def check_access(self, uid):
+        """
+        Validate access permission to the item
+        """
+        item = api.content.get(UID=uid)
+        if not item:
+            raise NotFound(self, uid)
+        if not api.user.has_permission(
+            "rer.customersatisfaction: Access Customer Satisfaction", obj=item
+        ):
+            raise Unauthorized(
+                _(
+                    "show_feedbacks_unauthorized",
+                    default=u"You don't have access to this content.",
+                )
+            )
+
     def get_data(self):
-        logger.info("QUI")
         tool = getUtility(ICustomerSatisfactionStore)
         query_vote = self.request.form.get("vote", "")
         uid = self.request.form.get("uid", "")
-        if not uid:
-            return {
-                "error": _(
-                    "show_feedbacks_missing_uid",
-                    default=u"You need to provide a UID.",
-                )
-            }
         query = {"uid": uid}
         search_results = tool.search(query=query)
         tot = len(search_results)
@@ -42,7 +66,6 @@ class View(BrowserView):
 
         if tot == 0:
             return res
-
         res["title"] = search_results[0]._attrs.get("title", "")
         for review in search_results:
             data = self.format_data(review)